refactor(tasks): move validator to authorizer package
